Task_07_02: auto merge of <Pull Request [#504]>, branch "a_kramarov"

+++ quadratic/views.py
@@ -3,7 +3,7 @@ import math
 from quadratic.forms import QuadraticForm
 
 def get_value(data, name):
-    return float(data.get(name))
+    return int(data.get(name))
 
 def discriminant(data):
